+++ src/main/java/match/MatchResults.java
@@ -1,4 +1,4 @@
-package result;
+package match;
 
 import java.util.LinkedHashMap;
 import java.util.List;
