+++ src/test/java/bet/BetResultsTest.java
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
 import player.Participant;
-import result.MatchResult;
+import match.MatchResult;
 
 class BetResultsTest {
